Moved getPage methods to trait. Added error log.
