+++ app/controllers/actions/Actions.scala
@@ -18,13 +18,9 @@ package controllers.actions
 
 import models.MovementReferenceNumber
 import models.requests.{DataRequest, OptionalDataRequest}
-import play.api.libs.json.Reads
-import play.api.mvc.Results.Redirect
-import play.api.mvc.{ActionBuilder, AnyContent, Result}
-import queries.Gettable
+import play.api.mvc.{ActionBuilder, AnyContent}
 
 import javax.inject.Inject
-import scala.concurrent.Future
 
 class Actions @Inject() (
   identifierAction: IdentifierAction,
@@ -37,19 +33,4 @@ class Actions @Inject() (
 
   def requireData(mrn: MovementReferenceNumber): ActionBuilder[DataRequest, AnyContent] =
     getData(mrn) andThen dataRequiredAction
-}
-
-object Actions {
-
-  def getPage[T](page: Gettable[T])(block: T => Result)(implicit request: DataRequest[_], rds: Reads[T]): Result =
-    getPage[T, Result](page)(block)(identity)
-
-  def getPageF[T](page: Gettable[T])(block: T => Future[Result])(implicit request: DataRequest[_], rds: Reads[T]): Future[Result] =
-    getPage[T, Future[Result]](page)(block)(Future.successful)
-
-  private def getPage[T, R](page: Gettable[T])(block: T => R)(f: Result => R)(implicit request: DataRequest[_], rds: Reads[T]): R =
-    request.userAnswers.get(page) match {
-      case Some(value) => block(value)
-      case None        => f(Redirect(controllers.routes.SessionExpiredController.onPageLoad()))
-    }
 }
